pull request amends p2
